backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
python-dotenv

# Additional dependencies for AI, Celery, and Redis
celery
redis
requests

# Development and Testing tools
django-debug-toolbar
pytest
pytest-django
